requirements.txt
openai
pandas
faiss-cpu
python-dotenv